Get rid of import cycles. Move K8s runtime into main runtime package
